rename list for available topics
